Step 0: mk/ Update Minix platform definition
We are transitionning from a full platform to a NetBSD variant,
per the project policy of having a NetBSD userland. There is then
little meaning in adapting all the NetBSD tests to also be true for
Minix.

This means the following is set as:
 * OPSYS = NetBSD
 * OS_VARIANT = Minix

This also has the consequence of obsoleting a few files in
pkgtools/bootstrap-mk-files.